Fix build error for Linux configure build of all-clusters-app

+++ src/app/clusters/thermostat-server/thermostat-server.cpp
@@ -408,7 +408,7 @@ uint8_t CountUpdatedPresetsAfterApplyingPendingPresets(Delegate * delegate)
         numberOfPendingPresets++;
     }
 
-    return numberOfPresets + numberOfPendingPresets - numberOfMatches;
+    return static_cast<uint8_t>(numberOfPresets + numberOfPendingPresets - numberOfMatches);
 }
 
 /**
